Fix the liquidation calculation
(fplus >= fminus) and (tz_mint >= tz_liq), which means that the
liquidation limit exceeds the burrowing limit. In turn, this means
that the burrow can be overburrowed, but NOT a candidate for
liquidation at the same time. In order to ensure that after
liquidation (if we end up selling tez for the expected price, that
is) the burrow is neither overburrowed, nor liquidatable, we must
target the burrowing limit during liquidation, NOT the liquidation
limit.