On liquidation
Deal with all possible scenaria:

Case A: the outstanding kit does not exceed the liquidation limit
so the request for liquidation is unwarranted.
==> Don't liquidate.
==> Result: non-liquidatable, (overburrowed/underburrowed, unknown!)

Case B: the burrow qualifies for liquidation, but we cannot even
refill the creation deposit.
==> Give the liquidation reward, auction the remaining collateral
    (<1 xtz) in its entirety, and close the burrow.
==> Result: closed

Case C: the burrow qualifies for liquidation, we can refill the
creation deposit, but with the current price it's impossible to
make the burrow not undercollateralized.
==> Give the liquidation reward, stash 1xtz creation deposit, and
    auction the remaining collateral in its entirety.
==> Result: overburrowed, liquidatable

Case D: the burrow qualifies for liquidation, and can be collateralized.
==> Give the liquidation reward, stash 1 xtz creation deposit, and
    auction exactly as much collateral as is needed to restore the
    burrow to a good state.
==> Result: underburrowed, non-liquidatable