Preliminary switch to Int64 for Kit, Tez, and FixedPoint
Note that due to precision issues, now liquidation does not ensure
that the burrow is not overburrowed (due to truncation, we sell less
tez than we should). This is only one of the many issues that we can
start fixing now.

Also note that the representation and the operations for Tez, Kit,
and FixedPoint are quickly drafted; we still have to decide on how
they should look internally.