uniswap.make_initial start with empty uniswap
Note that this can cause a little bit of trouble, because it means
that the initial price we get from uniswap is undefined (0/0).

Even after calling uniswap.add_liquidity for the first time though
we have the same problem, because we only query the price from the
last block (using uniswap.kit_in_tez_in_prev_block).

We solve this by making current and prevous price be identical after
uniswap.add_liquidity is called for the first time. After that, they
have a block difference, as expected.