avl: remove inconsistent parent change
Instead of inconsistently setting the parent of left in
Avl.ref_split to null in some cases but not in others,
let's just leave the change to Avl.take, to be performed
once, at the end. Alternatively, we could do this in
Avl.ref_split, but this increases bigmap reads and writes
significantly.