closer to ligo: ticket API
* Introduce missing tezos.mli
* Move contents of ticket.ml into tezos.ml
* Rename ticket API to match that of ligo:
  - Ticket.read => Tezos.read_ticket
  - Ticket.join => Tezos.join_tickets
  - Ticket.split => Tezos.split_ticket
  - Ticket.create => Tezos.create_ticket

Note that the operations are now properly named, but the type itself
not so. It should be `ticket`, not `Tezos.ticket`. The only way to
achieve that the way I see it is to put the type definition in
`ligo.mli` and the oprations in `tezos.mli`, but that exposes the
representation; not nice.