Move tez-related non-ligo functions out of ligo.mli
One of many. We have to do the same for all functions in ligo.mli
that are not provided by ligo; if we cannot do so, we have a problem.
I introduced common.ml as a destination for these functions so that
I can proceed, but we can move them anywhere we like (even inline
them) later on.