closer to ligo: fix input format for int, nat, and tez
This allows us to write arbitrarily big numbers, as we would in LIGO
(thus addressing the issues with defining the FixedPoint.scaling_factor).

We still bypass the "literals-only" expectation by using "string_of_int"
in the tests, but these shall never be translated to LIGO anyway.

Fixed also the 'sed' commands to allow for +/- in the front of ints and
to allow underscores within nats/ints/tez (LIGO allows this as well),
since we use quite some underscores in the OCaml code.