Introduce entrypoint %receiveLiquidationSlice
When processing a liquidation slice, the checker contract needs to
notify the corresponding burrow contract to send the actual tez
from the slice to the checker contract. Now the checker entrypoint
%receiveLiquidationSlice receives tez from burrows (only) and the
burrow entrypoint %burrowSendSliceToChecker emits the transaction
to checker.

I would have liked it much better if we reused %burrowSendTezTo,
and changed its type
  from BurrowSendTezTo : of (tez * address)
  to   BurrowSendTezTo : of (tez * unit contract)
but that introduces a module cycle right now; let's consider this
issue at some point.