Remove the delegate from the burrow representation
This is meta information that is not to be represented in checker.
Keeping it means (a) more data to store and process, and (b) even
bigger types than what we already have.

On it's own it brings a 77921 - 77586 = 335 bytes reduction, but it
might also allow us to uninline the touch_liquidation_slice fn, with
a massive effect.