wip: add Map.fold, and switch to list representation
Comment taken from the code:
Represent them like lists internally. This means that performance can be
horrible, but since we use them only for the oracles, this should not be a
problem. Why lists? Because Map.fold in LIGO works in increasing key order,
which we cannot ensure if we use the hash-based implementation of big_map;
the semantics would be wrong. List might be slow, but they are correct.