Revert Ligo.map (commit ea6fcfbc520570bafcd6476d1e6ae755b412d48a)
I took the need for Ligo.map to be de facto, but I was wrong. Given
the lack of homogeneity in the oracle APIs, I find it unlikely that
we can put them all in the same map (the entrypoints have not only
different names, but also different types).

Also, even if we could use a map for this, in LIGO Map.fold
traverses the key-value pairs in lexicographic order of keys, which
is impossible to achieve using a hash map.

If we really need a map implementation eventually, we shall write
a new one up, probably using something inefficient (e.g. list), but
correct.