Leave liquidation_reward_percentage a ratio (68529 => 68472)
instead of converting the ratio to a fixedpoint and then converting
the fixedpoint to a ratio again, which is what we used to do. This
change means that we sometimes generate a little more generous
liquidation reward (we still use flooring, ofc, but only once, at
the end), but it's clearer computationally, and easier to specify.