Refactor touch_liquidation_slice (65156 => 64912)
Touching a liquidation slice changes the state in 3 distinct ways:
(a) updates liquidation auction state (b) updates the state of the
burrow the slice refers to, and (c) updates the parameters to
account for the burned kit. The three are independent from each
other which means that we can perform them in any order (this ofc
applies to touch_liquidation_slices_rec as well), which in turn
_might_ allow us to move liquidation auctions into a separate
contract without too much effort after all. This small reordering
seems to drop the byte count a little too.