Add additional tests for burrow.ml (#61)
Added a new test module for burrow.ml and adds some preliminary unit tests for portions of burrow.ml which are not already tested in testLiquidation.ml. Also added a bunch of tests for the various functions in burrow.ml which should be checking that the burrow has been touched and fail otherwise as well as couple of property tests for checking that burrow_withdraw_tez and burrow_mint_kit throw errors in cases where they would cause the burrow to be overburrowed.