Update compute_tez_to_auction and burrow_request_liquidation
Updated compute_tez_to_auction to return an int instead of tez to
avoid cases where an exception would be thrown for results which
are negative and updated burrow_request_liquidation accordingly.

Also added tests which ensures that 1.) no exception is thrown in
compute_tez_to_auction if the result is negative and 2.) checks
that the values of the constants used in compute_tez_to_auction
do not break the implicit assumption in compute_tez_to_auction that
the constants are not reciprocals of eachother.