Revert 8dcecbd (i.e., replace some (||) with if-then-else)
Because currently LIGO and OCaml disagree on the semantics of (||):
OCaml short-circuits (magic), while LIGO evaluates both arguments.
We can probably revert again, if the semantics of LIGO change, but
that's certainly not for now.

NOTE: There are a few more uses of (||) and even more of (&&), but
I believe they are harmless, so I'm leaving them as they are for now.