[skip ci] Fix calculation of compute_min_kit_for_unwarranted
Now it rounds up only once, at the end of the calculations, being
consistent with the rest of the module.

NOTE: During the development of this I realized that we have not
tested this function at all. Even worse, I think there might be
legitimate cases that it fails to handle (when the collateral is
zero).