Initialize the outstanding and circulating kit in parameters at zero
Until now we'd start the outstanding_kit and circulating_kit in
parameters from 1mukit. The reason for that has been that touching
checker only multiplies them with indices so they'd effectively
stay at zero. However, this is not necessarily true: touching
checker creates an out-of-thin-air reward, effectively increasing
the circulating kit, and minting kit from burrows increases the
outstanding kit.