Add first pass at a list module for burrow liquidation slices
Added a first version of a module with utilities for operation on
the burrow slice list in the liquidation auction state. The current
idea is to use this API for inserting slices instead of operating
on the AVL + burrow_slices big map directly as is currently done
to allow for more easy testing.