Track the kit and lqt checker owns on the ledger
This includes kit stored in the cfmm, kit from winning bids for
liquidation auctions, and liquidity tokens stored in the cfmm.

This change allows us to check more invariants:
1. `total_kit_on_the_ledger = state.parameters.circulating_kit`
2. `checker's_kit_on_the_ledger + 1mukit >= state.cfmm.kit`
3. `checker's_lqt_on_the_ledger + 1n = state.cfmm.lqt`

NOTE 1: The `+1`s are needed because the cfmm keeps 1 phantom
token per category so that it never has to divide by zero.)

NOTE 2: The second assertion is an inequality because checker
also stores kit from bids.

This PR
* Extends `assert_checker_invariants` to additionally check
  the above three invariants.
* Adds a call to `assert_checker_invariants` at the beginning
  and end of every entrypoint (strict or lazy) and view.
* Adds extra assertions in the cfmm-related entrypoints when
  we know that the total amount of kit and/or lqt on the fa2
  ledger should be preserved or change by a fixed amount.

Because of the additional strictness I had to change several
tests; until now we'd frequently create input states that
violate some or all of these invariants.