Add big_map utility module to Python client library (#145)
Added a utility module to checker_client which provides helpers for working with big_maps in pytezos. Specifically this library includes functions for identifying keys in a contract's big_maps by scanning a chain for diffs and a supplementary method for rendering the contract's storage with the full content of the big_map.