Include checkerMain and checkerEntrypoints in the coverage report (#166)
I got a little confused by this one, since `checkerMain.ml` and
`checkerEntrypoints.ml` are both already listed in `src/dune` after
all. Apparently `bisect_ppx` is a little lazy:
having `src/checkerEntrypoints.ml` and `src/checkerMain.ml` around
is not enough, but the testsuite must have _some_ dependency on
them for them to be included on the report. Simply `open`ing them
doesn't do it either. So, this PR:
- Adds basic tests for `CheckerMain.initial_wrapper`, to ensure
  that `checkerMain.ml` and `checkerEntrypoints.ml` are included
  in the test coverage report.
- Makes `make test-coverage` have an explicit dependency on
  `src/checkerEntrypoints.ml`, since the file is now needed for
  instrumentation.