Add unit test for burrow_is_cancellation_warranted
Added a unit test for burrow_is_cancellation_warranted which covers
a case identified in mutation testing where changing the logic for
computing the kit in the equation within
burrow_is_cancellation_warranted did not break existing tests. This
test implementation is configured to be at the boundary between
true / false returns for burrow_is_cancellation_warranted and will
fail if the burrow's outstanding kit is increased by even a single
mukit.