Add call to FA2 balance_of entrypoint in end-to-end tests
Added a call to balance_of in the end-to-end test suite. Since
balance_of generally requires a contract to accept its callback,
this test uses pytezos' callback_view functionality which locally
interprets the entrypoint call and returns the emitted callback
operation. As such, this test doesn't provide an estimate of the
gas cost of balance_of but still provides a way to check 1.) whether
the entrypoint can successfully be called and 2.) the accurate of the
balance in the emitted callback operation.