Restore correct behavior for dune build
Following a334f05 `dune build @install` would not work correctly,
because src/dune did not specify a public name (so "checker" was
assumed), but the opam file in src was named "src.opam" so the
names did not match and `dune build @install` would silently do
nothing.

To address this, this commit adds the public name "checker" to
dune/src and the accompanying (currently empty) opam file at the
top-level.