Additional profiling in end to end tests (#201)
Added additional profiling logic in our end-to-end tests for recording gas costs and other related metadata during the liquidation auction stress test. Similar to other statistics, this data can be written out to a JSON file by specifying an environment variable. This data can then be plotted with the new e2e/plot-gas-profiles.py script which is also used to generate profile plots as CI artifacts. Note that the CI time is pretty dramatically effected by the number of liquidation slices we cancel, so if the CI time becomes too much we can reduce this value (at the cost of samples for plotting).