Refactor Earthfile
Refactored Earthfile to use a common base image for both ocaml
and python project dependencies to facilitate the creation of
a dev container image. Since in terms of final artifacts we are
primarily interested in the contract code and CLI continer, having
a larger base image doesn't really effect the final image size.

Also updated the README (along with the formatting similar to what @gkaracha
had applied) and added some extra stages to the Earthfile.