Paste-and-change, for now. [Gotta fix this]
The problem is that we have checker-specific logic in fa2Interface.ml
so we cannot really reuse it for the tezWrapper (e.g., the token_ids
used are checker-specific, no vaults are involved, ..). Ideally, I'd
keep fa2Interface.ml clean (i.e., leave only the types and basic
infrastructure in it), and have a separate module (e.g., fa2State or
something) with checker-specific data. This way, tezWrapper can open
only fa2Interface and implement the specialized functions that it
needs.