Add some tests for src/ptr.ml
Two of those (test_ptr_null_is_zero and test_ptr_next_difference)
are tied to our implementation of ptr, and the other one is the
most important property to have (test_ptr_next_injective).

The tight tests are used primarily to catch untested mutations.

Also remove Ptr.ptr_init since it is not used at all in src/ (though
it was used a little in tests/). This reduces the surface area to
mutate.