Switch to FA2 for collateral, exclusively (#269)
First part of #213: integration of the tez wrapper into checker's
workflow. In essence this PR makes checker deal with FA2
exclusively. To keep CI green, usage of the tez wrapper is
hardwired; adding the switch to be able to change the collateral
type should be part of a separate PR. This PR is already pretty
big.

Other changes:
* Removed the restrictions on `from_` and `to_` in
  `TezWrapper.transfer` to save on gas costs (they were optional
  anyway).
* Fixed a bug in `TezWrapper.transfer` (transfer of zero tez
  should not fail due to non-originated vaults).
* Made `tezWrapper.ml` more verbose to significantly reduce its
  gas costs (see bot message below).